ISSUE #3396 feedback addressed
